Day02
Again on the simpler side. I opted not for a regex (I couldn't think on the top of my head on how to do the repetitive groups) and chose to do simple string splits, which makes the code rather simple and explicit.

On part 1 one could then early return on seeing the maximum. On part 2 one had to keep a running maximum of seen values.