Day03
A bit on the trickier side for a day 3, but it's Sunday after all.
I opted for an early point class (record) with the usual `neighbors`, as well as `left` and `right` operations to move around in a point-map/grid.

Part 1 was then mostly just parsing the grid and then getting a more or less clever way to find a number from a number-candidate in the grid, without finding the whole number twice (using a "seen" set). I was first worried, that a number part could be connected to multiple numbers, but by inspecting the input I saw, that Eric was not (yet) that mean to us.

Part 2 then uses mostly the same logic, but we select a subset of all symbols (super easy with the stream-filter on the `entrySet` approach) and do only add up the gear-ratios when we actually have two ratios.

All in all a cool and not too hard Sunday Day 03.