Day 04
So for part 1 and 2 you need to know the method `List::retainAll` (or equivalent in your language) to get the overlaps of one list/set to another. Doing this by hand is not hard and can be done in linear runtime (k*n, k is the size of the smaller lookup, n of the array).

For part 1 one needed then to see that you could calculate the value directly by using the powers of 2.

In part 2 you had to watch out not to do the calculations exponentially but calculate the sums of each card directly. Which is not that hard to see: you always add the current card amount to the following cards. Eric was even so nice and not include an overflow scenario, the cards line up (you do not win more cards on the last few cards as you actually have).