Day06
Compared to yesterday, this was smooth sailing. Part 1 was simple, and you just calculate all the races separately to get all possibilities.
In part 2 I first hesitated to do it the brute-force way, but as I saw the final numbers (in my case a time of `54708275` which would be the loop size and distance of `239114212951253` which would only be used for comparison), I immediately knew it would work w/o some math tricks.