Day09
A good day to play around with the Java-Stream-API. As usual, the first step was to parse the input, today in simple list of longs. We then had to create a "repeat finding the difference until all values are 0" function, which - when not implemented too naively - can be a brute force implementation for part 1 and part 2. One thing to look out for is the termination check (all numbers 0) and doing that efficiently (i.e. using `anyMatch(value->value!=0)` is fast and stops at the first non 0 value).

For part 1 we then just sum up the last values of that list of differences.

For part 2 we have to be a bit clever, first we need the first elements and in the right order (note that they are added in reverse order in the `differencesUntilZeroes` method, because in part 1 it does not matter and part 2 we save one reversal of the stream/list). And then we just "subtracting the running total from each value", a.k.a. a reduce with a subtraction starting at 0.

Not too complicated today, only about 13 minutes for part 2.