Day10
A loopy day. A lot of boilerplate overhead today with enums and directions, which I partially also integrated in the point class (for navigation in a grid, separating directions and points makes sense). So for part 1 it was mostly again a parsing exercise, where you actually would not need the loop, just a counter.
For part 2 I first played around with the idea of flood fill and some sort of "zoom the map" method to create room between the pipes as squeezing was allowed. However, a scanline approach where we just went from top to bottom and left to right seemed feasible. One had just to group the right corner peaces together.