Day 20 Part 1 was simple parsing and for me easily solvable with a queue approach. I struggled a bit with the difference of a Module and a Signal, but separating them each into their record helped unclutter the logic quiet a bit.
Part 2 then was really hard.
I first let a brute-force approach run, just in case it would work after some minutes of running. No dice. So I had to do, what I usually don't really like to do in AoC: inspect the input for your specific case. I always use this as last resort, but this time Eric was nice to us: you can easily see, that having not too many instructions, this probably explodes on the last few before our target "rx". I then realized, that all sources of the conjunction rx are conjunctions as well. So I first tested in an endless loop, when one of them would be set to false (low-signal), and see there: a few thousand steps. I then checked one more candidate and as well: a few (different) thousand steps it went to false. I then checked if the steps repeated in cycles and they did immediately.
From this point onward it was straight forward: extract the actual process into a helper method, search for the "cycles" of each of the conjunction and LCM them all.